requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
email_validator
faster-whisper
huggingface-hub
python-docx
llama-cpp-python
